FB grabber Update
FB Grabber updated to ignore equipment listings that are were not listed on facebook when considering duplicates for deletion.

If item was listed through facebook and it is encountered by the script collecting data from facebook, it will be removed from the database, if the item was not added through facebook to begin with, it will not be deleted.